services/core/requirements.txt
graphql-core==3.2.5
PyJWT==2.8.0
setuptools==75.5.0
